refactor(user_center): 更改用户 ID 类型为 BigDecimal，更新相关方法及接口参数

+++ user_center/src/main/java/io/github/lionheartlattice/user_center/controller/UserController.java
@@ -40,7 +40,7 @@ public class UserController {
     @Operation(summary = "获取用户详情", description = "根据用户 ID 获取用户的详细信息")
     @PostMapping("detail")
     public ApiResult<UserUpdateDTO> getById(
-            @RequestParam Long id) {
+            @RequestParam BigDecimal id) {
         return ApiResult.success(userService.getById(id));
     }
